Fix failures in iOS 9
There appear to be some changes in the recent Rubymotion updates (I
haven't been able to find where) that means some specs were failing when
run with iOS 9 as the target.

This change introduces a fix specifically to the `Constants.get` method
as used by `UIActivityViewController`.

The error was, passing activities to exclude from the activity view
caused an error where an Array was expected but the splat wasn't
creating one.

The other issue this fixes is to delegate the `tag` property on a
`UINavigationViewItem` to it's child `UIBarButtonItem` in the test
implementation (a `UINavigationViewItem` doesn't have a `tag` property).